+++ cmd/worker/internal/migrations/migrators/extsvc_webhook_migrator_test.go
@@ -1,4 +1,4 @@
-package oobmigrators
+package migrators
 
 import (
 	"context"